+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Settings: Added a hint under Project Identifier explaining where to adjust background image position and opacity.